+++ pymatviz/utils/__init__.py
@@ -4,11 +4,6 @@ from __future__ import annotations
 
 import os
 
-
-PKG_DIR: str = os.path.dirname(__file__)
-ROOT: str = os.path.dirname(PKG_DIR)
-
-
 from pymatviz.utils.data import (
     atomic_numbers,
     bin_df_cols,
@@ -48,3 +43,7 @@ from pymatviz.utils.typing import (
     R,
     T,
 )
+
+
+PKG_DIR: str = os.path.dirname(__file__)
+ROOT: str = os.path.dirname(PKG_DIR)
